test: 인수 테스트 통합 및 LineSectionRestAssuredUtils 분리
